* rads-0_ascii_rC-env:
  Release C foundations. New products (Geosat, RADS). RADS service. Export ASCII. Remote Counter.

# Solved Conflicts:
#	projects/qtc/brat-v4-subdirs.pro
#	projects/qtc/tests/testTools/target_wrapper.sh
#	source/common/ApplicationStaticPaths.cpp
#	source/common/ApplicationStaticPaths.h
#	source/libbrathl/ExternalFilesFactory.cpp
#	source/libbrathl/NetCDFFiles.cpp
#	source/libbrathl/ProductGeosatGDR.h
#	source/new-gui/brat/ApplicationPaths.cpp
#	source/new-gui/brat/ApplicationSettingsDlg.cpp
#	source/new-gui/brat/ApplicationSettingsDlg.h
#	source/new-gui/brat/BratApplication.h
#	source/new-gui/brat/BratMainWindow.cpp
#	source/new-gui/brat/BratMainWindow.h

+++ libbrathl/BratAlgoFilterKernel.cpp
@@ -21,8 +21,8 @@
 #include "brathl.h" 
 
 
-#include "new-gui/Common/tools/TraceLog.h" 
-#include "new-gui/Common/tools/Exception.h" 
+#include "common/tools/TraceLog.h"
+#include "common/tools/Exception.h"
 #include "Expression.h" 
 #include "Product.h" 
 #include "ProductNetCdf.h" 
@@ -606,7 +606,7 @@ void CBratAlgoFilterKernel::PrepareDataWindow2DSymmetrically(CExpressionValue& e
 }
 
 //----------------------------------------
-void CBratAlgoFilterKernel::PrepareDataWindow2DSymmetrically(double* dataValue, uint32_t nbValues, uint32_t shiftSymmetry /* = 0 */)
+void CBratAlgoFilterKernel::PrepareDataWindow2DSymmetrically(double* dataValue, size_t nbValues, uint32_t shiftSymmetry /* = 0 */)
 {
     UNUSED(nbValues);
